fix(appliance): allow services to be opted into
Creating a SourcegraphSpec in the web app postSetupHandler with
Spec.Jaeger.Disabled=false (by hacking it in as an experiment) didn't
actually override the default of true set in defaults.go, so we couldn't
actually deploy jaeger in practice.

This was probably an interaction of unmarshalling a bool with omitempty,
that appears to be fixed by using a pointer.